refactor(runtime): route/ package, explicit line_routing threading over ContextVar
