+++ python/mirage/runtime/route/__init__.py
@@ -1,0 +1,33 @@
+# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+from mirage.runtime.route.decide import (decide_line, evaluate_route,
+                                         evaluate_script)
+from mirage.runtime.route.errors import RoutingDecisionError
+from mirage.runtime.route.facts import command_facts
+from mirage.runtime.route.types import (CommandFacts, LineRouting,
+                                        RouteContext, RouteFn, RouteScript)
+
+__all__ = [
+    "CommandFacts",
+    "LineRouting",
+    "RouteContext",
+    "RouteFn",
+    "RouteScript",
+    "RoutingDecisionError",
+    "command_facts",
+    "decide_line",
+    "evaluate_route",
+    "evaluate_script",
+]
