+++ python/mirage/runtime/route/decide.py
@@ -14,169 +14,14 @@
 
 import asyncio
 import inspect
-from collections.abc import Awaitable, Callable
-from dataclasses import dataclass, field
+from collections.abc import Callable
 from typing import Any
 
-import tree_sitter
-
-from mirage.commands.spec import SPECS
 from mirage.runtime.base import Runtime
 from mirage.runtime.python.monty import _MirageOS, pydantic_monty
+from mirage.runtime.route.types import (LineRouting, RouteContext, RouteFn,
+                                        RouteScript)
 from mirage.runtime.table import VfsEntry, bind_commands, pin_bindings
-from mirage.shell.types import NodeType
-
-
-class RoutingDecisionError(ValueError):
-    """A pin, route, or script could not decide the line.
-
-    Raised for caller-fixable routing mistakes (unknown pin name, a
-    script that does not parse, a missing monty extra) so execute()
-    propagates them loud instead of folding them into the line's
-    IOResult like a command failure.
-    """
-
-
-_WORD_TYPES = (NodeType.COMMAND_NAME, NodeType.WORD, NodeType.STRING,
-               NodeType.RAW_STRING, NodeType.NUMBER, NodeType.CONCATENATION)
-
-
-@dataclass(frozen=True, slots=True)
-class CommandFacts:
-    """Parse facts for one command of the line being routed.
-
-    Args:
-        command (str): the command name (first word).
-        words (tuple[str, ...]): every word of the command, name first.
-        known (bool): whether the command has a builtin spec.
-        paths (tuple[str, ...]): absolute-path operands.
-    """
-
-    command: str
-    words: tuple[str, ...]
-    known: bool
-    paths: tuple[str, ...]
-
-
-@dataclass(frozen=True, slots=True)
-class RouteContext:
-    """Facts about the line being routed, parse-before-route.
-
-    Args:
-        line (str): the raw command line.
-        commands (tuple[CommandFacts, ...]): parsed commands, empty on
-            a syntax error.
-        command (str): the first command name, "" when unparsable.
-        known (bool): whether the first command has a builtin spec.
-        cwd (str): session working directory.
-        env (dict[str, str]): session environment.
-        session_id (str): session hosting the line.
-        agent_id (str): agent executing the line.
-        mounts (tuple[str, ...]): workspace mount prefixes.
-    """
-
-    line: str
-    commands: tuple[CommandFacts, ...]
-    command: str
-    known: bool
-    cwd: str
-    env: dict[str, str]
-    session_id: str
-    agent_id: str
-    mounts: tuple[str, ...]
-
-    def to_dict(self, runtime: Runtime | None = None) -> dict[str, Any]:
-        """The monty-facing ctx payload.
-
-        Args:
-            runtime (Runtime | None): the runtime being asked, added as
-                ctx["runtime"] for per-runtime scripts.
-        """
-        payload: dict[str, Any] = {
-            "line":
-            self.line,
-            "commands": [{
-                "command": c.command,
-                "words": list(c.words),
-                "known": c.known,
-                "paths": list(c.paths),
-            } for c in self.commands],
-            "command":
-            self.command,
-            "known":
-            self.known,
-            "cwd":
-            self.cwd,
-            "env":
-            dict(self.env),
-            "session_id":
-            self.session_id,
-            "agent_id":
-            self.agent_id,
-            "mounts":
-            list(self.mounts),
-        }
-        if runtime is not None:
-            payload["runtime"] = {
-                "name": runtime.name,
-                "captures": list(runtime.captures),
-            }
-        return payload
-
-
-# A per-runtime willingness script: a callable on the RouteContext
-# returning a truthy verdict, or monty source whose last expression is
-# the verdict. Mirrors the TS RouteScript.
-RouteScript = Callable[[RouteContext], bool | Awaitable[bool]] | str
-
-# The global route: a callable on the RouteContext returning a runtime
-# name (or None to pass), or monty source. Mirrors the TS RouteFn.
-RouteFn = Callable[[RouteContext], str | None | Awaitable[str | None]] | str
-
-
-@dataclass(frozen=True, slots=True)
-class LineRouting:
-    """The one-line placement decision the dispatcher consults.
-
-    Args:
-        bindings (dict[str, Runtime]): command -> runtime for this
-            line.
-        vfs_allowed (bool): whether unbound commands may run on the
-            vfs executor; False turns them into admission failures.
-        captured (frozenset[str]): commands captured by some entry;
-            an unbound captured command is an admission failure (its
-            capturers all refused), never a silent fallback.
-    """
-
-    bindings: dict[str, Runtime] = field(default_factory=dict)
-    vfs_allowed: bool = True
-    captured: frozenset[str] = frozenset()
-
-
-def command_facts(ast: tree_sitter.Node) -> tuple[CommandFacts, ...]:
-    """Extract per-command parse facts from a parsed line.
-
-    Args:
-        ast (tree_sitter.Node): the parsed tree-sitter root node.
-    """
-    facts: list[CommandFacts] = []
-    stack = [ast]
-    while stack:
-        node = stack.pop()
-        if node.type == "command":
-            words = tuple(
-                child.text.decode() for child in node.children
-                if child.type in _WORD_TYPES and child.text is not None)
-            if words:
-                facts.append(
-                    CommandFacts(
-                        command=words[0],
-                        words=words,
-                        known=words[0] in SPECS,
-                        paths=tuple(w for w in words[1:] if w.startswith("/")),
-                    ))
-        stack.extend(reversed(node.children))
-    return tuple(facts)
 
 
 async def _eval_monty(source: str, ctx_payload: dict[str, Any],
