+++ python/mirage/workspace/executor/command.py
@@ -32,7 +32,7 @@ from mirage.io import IOResult
 from mirage.io.stream import async_chain, materialize, wrap_cachable_streams
 from mirage.io.types import ByteSource
 from mirage.runtime.base import Runtime
-from mirage.runtime.pin import current_line_routing
+from mirage.runtime.route import LineRouting
 from mirage.shell.call_stack import CallStack
 from mirage.shell.job_table import JobTable
 from mirage.shell.types import ERREXIT_EXEMPT_TYPES
@@ -143,21 +143,21 @@ def _check_mount_root_guard_raw(
 
 
 def _line_runtime(
-        cmd_name: str,
-        registry: MountRegistry) -> tuple[Runtime | None, IOResult | None]:
+        cmd_name: str, registry: MountRegistry,
+        routing: LineRouting | None) -> tuple[Runtime | None, IOResult | None]:
     """Resolve a command against the line's routing decision.
 
-    With no decision active, the workspace's static bindings apply.
-    With one, the line's bindings win; an unbound command whose
-    capturers all refused, or any unbound command when the vfs rung
-    was refused, is an admission failure: exit 126, "no runtime
-    accepted this line", like a shell refusing to exec.
+    With no decision, the workspace's static bindings apply. With one,
+    the line's bindings win; an unbound command whose capturers all
+    refused, or any unbound command when the vfs rung was refused, is
+    an admission failure: exit 126, "no runtime accepted this line",
+    like a shell refusing to exec.
 
     Args:
         cmd_name (str): the command being dispatched.
         registry (MountRegistry): registry holding static bindings.
+        routing (LineRouting | None): the typed line's decision.
     """
-    routing = current_line_routing()
     if routing is None:
         return registry.runtime_bindings.get(cmd_name), None
     runtime = routing.bindings.get(cmd_name)
@@ -203,6 +203,7 @@ async def run_on_mount(
     stdin: ByteSource | None = None,
     resolve_hint: PathSpec | None = None,
     mount: MountEntry | None = None,
+    line_routing: LineRouting | None = None,
 ) -> tuple[ByteSource | None, IOResult]:
     """Run one already-parsed command on the mount that owns its paths.
 
@@ -257,7 +258,7 @@ async def run_on_mount(
     stat_overlay = (functools.partial(_namespace_stat_overlay, namespace)
                     if cmd_name == "ls" and namespace is not None else None)
 
-    line_runtime, denial = _line_runtime(cmd_name, registry)
+    line_runtime, denial = _line_runtime(cmd_name, registry, line_routing)
     if denial is not None:
         return None, denial
 
@@ -462,6 +463,7 @@ async def handle_command(
     call_stack: CallStack | None = None,
     job_table: JobTable | None = None,
     namespace: Namespace | None = None,
+    line_routing: LineRouting | None = None,
 ) -> tuple[ByteSource | None, IOResult, ExecutionNode]:
     """Execute a simple command.
 
@@ -588,8 +590,12 @@ async def handle_command(
                                       command=cmd_str,
                                       exit_code=code,
                                       stderr=refusal_msg)
-        run_single = functools.partial(run_on_mount, registry, session,
-                                       dispatch, namespace)
+        run_single = functools.partial(run_on_mount,
+                                       registry,
+                                       session,
+                                       dispatch,
+                                       namespace,
+                                       line_routing=line_routing)
         stdout, io = await handle_cross_mount(cmd_name,
                                               path_scopes,
                                               cross_texts,
@@ -706,7 +712,8 @@ async def handle_command(
                                     texts,
                                     flag_kwargs,
                                     stdin=stdin,
-                                    mount=mount)
+                                    mount=mount,
+                                    line_routing=line_routing)
 
     if warn_bytes:
         existing = await materialize(io.stderr) if io.stderr else b""
