+++ python/mirage/workspace/node/run_tree.py
@@ -18,6 +18,7 @@ from typing import Any, Callable
 from mirage.commands.builtin.utils.safeguard import apply_safeguard
 from mirage.io import IOResult
 from mirage.io.types import materialize
+from mirage.runtime.route import LineRouting
 from mirage.shell.barrier import BarrierPolicy, apply_barrier
 from mirage.shell.job_table import JobTable
 from mirage.workspace.mount import MountRegistry
@@ -38,6 +39,7 @@ async def run_command_tree(
     session: Session,
     stdin: Any,
     cancel: asyncio.Event | None,
+    line_routing: LineRouting | None = None,
 ) -> tuple[IOResult, ExecutionNode]:
     """Run a parsed command tree and finalize its output stream.
 
@@ -61,6 +63,9 @@ async def run_command_tree(
         session (Session): shell session state.
         stdin (Any): input stream.
         cancel (asyncio.Event | None): event used to abort mid-flight.
+        line_routing (LineRouting | None): the typed line's routing
+            decision, threaded to every command dispatch; None runs on
+            the static bindings.
 
     Returns:
         tuple[IOResult, ExecutionNode]: the finalized result (with
@@ -78,6 +83,7 @@ async def run_command_tree(
         session,
         stdin,
         cancel=cancel,
+        line_routing=line_routing,
     )
     stdout = await apply_barrier(stdout, io, BarrierPolicy.VALUE)
     if io.safeguard is not None and stdout is not None:
