+++ typescript/packages/core/src/workspace/executor/command.ts
@@ -32,7 +32,7 @@ import { mergeOverlayStat } from '../mount/namespace/overlay.ts'
 import { MountCommandUnsupported, type MountRegistry } from '../mount/registry.ts'
 import { Consumer, JOB_BUILTINS, route } from '../route/index.ts'
 import type { Runtime } from './runtime.ts'
-import type { LineRouting } from './route.ts'
+import type { LineRouting } from './route/index.ts'
 import type { Session } from '../session/session.ts'
 import { ExecutionNode } from '../types.ts'
 import { asyncChain } from '../../io/stream.ts'
