+++ typescript/packages/core/src/workspace/executor/route/errors.ts
@@ -1,0 +1,26 @@
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+/**
+ * A pin, route, or script could not decide the line. Caller-fixable
+ * routing mistakes (unknown pin name, a script that does not parse, a
+ * missing monty package) propagate loud instead of folding into the
+ * line's IOResult like a command failure.
+ */
+export class RoutingDecisionError extends Error {
+  constructor(message: string, options?: { cause?: unknown }) {
+    super(message, options)
+    this.name = 'RoutingDecisionError'
+  }
+}
