+++ typescript/packages/core/src/workspace/executor/route/index.ts
@@ -1,0 +1,18 @@
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+export { decideLine } from './decide.ts'
+export { RoutingDecisionError } from './errors.ts'
+export { commandFacts } from './facts.ts'
+export type { CommandFacts, LineRouting, RouteContext, RouteFn, RouteScript } from './types.ts'
