+++ typescript/packages/core/src/workspace/runtime_pin.test.ts
@@ -14,7 +14,7 @@
 
 import { describe, expect, it } from 'vitest'
 import { VfsEntry, type RunArgs, type Runtime, type RunResult } from './executor/runtime.ts'
-import type { RouteScript } from './executor/route.ts'
+import type { RouteScript } from './executor/route/index.ts'
 import { getTestParser } from './fixtures/workspace_fixture.ts'
 import { RAMResource } from '../resource/ram/ram.ts'
 import { MountMode } from '../types.ts'
